refactor : category 순수 jpa를 spring data jpa로 리펙토링
