+++ kumoh_board_backend/kumoh_board_backend/src/main/java/com/creativedesignproject/kumoh_board_backend/category/domain/entity/Category.java
@@ -1,4 +1,4 @@
-package com.creativedesignproject.kumoh_board_backend.category.domain;
+package com.creativedesignproject.kumoh_board_backend.category.domain.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
